Работает добавление песни, но не работает проигрывание

+++ app/src/main/java/com/example/tiget/musicplayer/ui/Library/LibSong.java
@@ -3,7 +3,7 @@ package com.example.tiget.musicplayer.ui.Library;
 import java.io.Serializable;
 
 public class LibSong implements Serializable {
-    public long id;
+    public int id;
     public String type;
     public String AuthorName;
     public String SongName;
@@ -11,7 +11,7 @@ public class LibSong implements Serializable {
     public int SongPreview;
     boolean explicit;
 
-    public LibSong(long id, String type, String AuthorName, String SongName, String SongUri, int SongPreview, boolean explicit){
+    public LibSong(int id, String type, String AuthorName, String SongName, String SongUri, int SongPreview, boolean explicit){
         this.id = id;
         this.type = type;
         this.AuthorName = AuthorName;
@@ -23,7 +23,7 @@ public class LibSong implements Serializable {
     }
 
 
-    public long getId(){
+    public int getId(){
         return id;
     }
